refactor(gms): reorganize GMS factory namespace
Fixes https://github.com/linkedin/datahub/issues/1658
