+++ gms/factories/src/main/java/com/linkedin/gms/factory/common/LocalEbeanServerConfigFactory.java
@@ -1,4 +1,4 @@
-package com.linkedin.common.factory;
+package com.linkedin.gms.factory.common;
 
 import io.ebean.config.ServerConfig;
 import io.ebean.datasource.DataSourceConfig;
