+++ gms/factories/src/main/java/com/linkedin/gms/factory/common/TopicConventionFactory.java
@@ -1,4 +1,4 @@
-package com.linkedin.common.factory;
+package com.linkedin.gms.factory.common;
 
 import com.linkedin.mxe.TopicConvention;
 import com.linkedin.mxe.TopicConventionImpl;
