+++ gms/factories/src/main/java/com/linkedin/gms/factory/dataset/DatasetSearchDAOFactory.java
@@ -1,8 +1,8 @@
-package com.linkedin.identity.factory;
+package com.linkedin.gms.factory.dataset;
 
-import com.linkedin.metadata.configs.CorpUserSearchConfig;
+import com.linkedin.metadata.configs.DatasetSearchConfig;
 import com.linkedin.metadata.dao.search.ESSearchDAO;
-import com.linkedin.metadata.search.CorpUserInfoDocument;
+import com.linkedin.metadata.search.DatasetDocument;
 import org.elasticsearch.client.RestHighLevelClient;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
@@ -12,16 +12,17 @@ import org.springframework.context.annotation.DependsOn;
 
 import javax.annotation.Nonnull;
 
+
 @Configuration
-public class CorpUserSearchDaoFactory {
+public class DatasetSearchDAOFactory {
   @Autowired
   ApplicationContext applicationContext;
 
-  @Bean(name = "corpUserSearchDAO")
-  @DependsOn({"elasticSearchRestHighLevelClient"})
   @Nonnull
+  @DependsOn({"elasticSearchRestHighLevelClient"})
+  @Bean(name = "datasetSearchDao")
   protected ESSearchDAO createInstance() {
-    return new ESSearchDAO(applicationContext.getBean(RestHighLevelClient.class), CorpUserInfoDocument.class,
-        new CorpUserSearchConfig());
+    return new ESSearchDAO(applicationContext.getBean(RestHighLevelClient.class), DatasetDocument.class,
+        new DatasetSearchConfig());
   }
 }
