+++ gms/factories/src/main/java/com/linkedin/gms/factory/identity/CorpUserDaoFactory.java
@@ -1,11 +1,11 @@
-package com.linkedin.identity.factory;
+package com.linkedin.gms.factory.identity;
 
-import com.linkedin.common.factory.TopicConventionFactory;
-import com.linkedin.common.urn.CorpGroupUrn;
-import com.linkedin.metadata.aspect.CorpGroupAspect;
+import com.linkedin.gms.factory.common.TopicConventionFactory;
+import com.linkedin.common.urn.CorpuserUrn;
+import com.linkedin.metadata.aspect.CorpUserAspect;
 import com.linkedin.metadata.dao.EbeanLocalDAO;
 import com.linkedin.metadata.dao.producer.KafkaMetadataEventProducer;
-import com.linkedin.metadata.snapshot.CorpGroupSnapshot;
+import com.linkedin.metadata.snapshot.CorpUserSnapshot;
 import com.linkedin.mxe.TopicConvention;
 import io.ebean.config.ServerConfig;
 import org.apache.kafka.clients.producer.Producer;
@@ -19,18 +19,18 @@ import javax.annotation.Nonnull;
 
 
 @Configuration
-public class CorpGroupDaoFactory {
+public class CorpUserDaoFactory {
   @Autowired
   ApplicationContext applicationContext;
 
-  @Bean(name = "corpGroupDao")
+  @Bean(name = "corpUserDao")
   @DependsOn({"gmsEbeanServiceConfig", "kafkaEventProducer", TopicConventionFactory.TOPIC_CONVENTION_BEAN})
   @Nonnull
   protected EbeanLocalDAO createInstance() {
-    KafkaMetadataEventProducer<CorpGroupSnapshot, CorpGroupAspect, CorpGroupUrn> producer =
-        new KafkaMetadataEventProducer(CorpGroupSnapshot.class, CorpGroupAspect.class,
+    KafkaMetadataEventProducer<CorpUserSnapshot, CorpUserAspect, CorpuserUrn> producer =
+        new KafkaMetadataEventProducer(CorpUserSnapshot.class, CorpUserAspect.class,
             applicationContext.getBean(Producer.class), applicationContext.getBean(TopicConvention.class));
-    return new EbeanLocalDAO<>(CorpGroupAspect.class, producer, applicationContext.getBean(ServerConfig.class),
-        CorpGroupUrn.class);
+    return new EbeanLocalDAO<>(CorpUserAspect.class, producer, applicationContext.getBean(ServerConfig.class),
+        CorpuserUrn.class);
   }
 }
