+++ gms/war/src/main/webapp/WEB-INF/beans.xml
@@ -4,7 +4,7 @@
        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-3.0.xsd"
        default-lazy-init="true">
 
-    <context:component-scan base-package="com.linkedin.common.factory,com.linkedin.dataset.factory,com.linkedin.identity.factory,com.linkedin.dataprocess.factory" />
+    <context:component-scan base-package="com.linkedin.gms.factory" />
 
     <!-- An InjectResourceFactory must be created.  It is ApplicationContextAware so
     must be instantiated via spring and then passed into rest.li -->
